LayoutTests/compositing/repaint/sticky-repaint-on-scroll-expected.txt

A
A
A
PASS repaintCount is 1
PASS successfullyParsed is true

TEST COMPLETE
